[hash-based-randomization] clean up story, move functions to separate util file
