+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/radio-widget.ts
@@ -14,6 +14,7 @@ import {defaulted} from "../general-purpose-parsers/defaulted";
 import {versionedWidgetOptions} from "./versioned-widget-options";
 import {parseWidgetWithVersion} from "./widget";
 import {parseWidgetsMap} from "./widgets-map";
+import { normalizeContent, simpleHash } from "../../utils/content-hash";
 
 import type {ParsedValue} from "../parser-types";
 
@@ -162,36 +163,6 @@ const parseRadioWidgetV0 = parseWidgetWithVersion(
 );
 
 // migrate functions
-
-/**
- * Simple hash function to create a short hash from a string
- * Uses a more robust algorithm for better distribution
- */
-function simpleHash(str: string): string {
-    let hash = 5381; // Use djb2 hash algorithm
-    if (str.length === 0) {
-        return "empty";
-    }
-
-    for (let i = 0; i < str.length; i++) {
-        const char = str.charCodeAt(i);
-        hash = (hash << 5) + hash + char; // hash * 33 + char
-    }
-
-    return Math.abs(hash).toString(36).substring(0, 8); // Base36 for shorter string, 8 chars for better uniqueness
-}
-
-/**
- * Create a normalized version of content for hashing
- * This removes extra whitespace and normalizes the text to ensure consistent hashing
- */
-function normalizeContent(content: string): string {
-    return content
-        .trim()
-        .replace(/\s+/g, " ") // Replace multiple whitespace with single space
-        .toLowerCase(); // Normalize case for consistency
-}
-
 export function migrateV3ToV4(
     widget: ParsedValue<typeof parseRadioWidgetV3>,
 ): ParsedValue<typeof parseRadioWidgetV4> {
